+++ eemeter/io.py
@@ -85,7 +85,7 @@ def meter_data_from_csv(
 
     # for pandas<0.24, which doesn't localize even with utc=True
     if df.index.tz is None:
-        df.index = df.index.tz_localize("UTC")
+        df.index = df.index.tz_localize("UTC") # pragma: no cover
 
     if tz is not None:
         df = df.tz_convert(tz)
@@ -152,7 +152,7 @@ def temperature_data_from_csv(
 
     # for pandas<0.24, which doesn't localize even with utc=True
     if df.index.tz is None:
-        df.index = df.index.tz_localize("UTC")
+        df.index = df.index.tz_localize("UTC") # pragma: no cover
 
     if tz is not None:
         df = df.tz_convert(tz)