+++ CHANGELOG.md
@@ -4,6 +4,12 @@ Changelog
 Development
 -----------
 
+* Placeholder
+
+2.10.0
+------
+
+
 * Add additional terms into ModelMetrics() class which can be used in fractional savings uncertainy computations.
 
 2.9.2